requirements.txt
Flask>=3.0,<4.0
python-dotenv>=1.0,<2.0
google-generativeai>=0.4,<1.0
gTTS>=2.5,<3.0